Begining of getting latest available jobs
Have initialized the work of getting dynamic information of available jobs, only added some dynamic data: look in list-of-adds.js. Have added company_name and job_title to first job but need to add dynamic data for the rest. The idea is to get the ten latest jobs from Arbetsformedlingen.